Better layer window and automatic mask generation
Added a new layer management window that lets the user delete layers,
and improved the UI of the add layer window. This improved UI makes it
explicit that a mask is not required, and shows a preview of what the
file + mask will look like on a sample backdrop.

This new UI has support for automatically generating a mask for layers
with transparency. This mask is created by running a breadth-first
search from each corner of the image looking for transparent pixels.
This is done to avoid masking out transparent enclosed parts of the
underlying image since enclosed holes in a layer-image is where the
underlying image is supposed to be shown.