Suggest the last-used path in the save-dialog
We now keep track of the directory and filename that was used the last
time we saved a token. When opening the save-dialog, it will open to the
same directory as last time, with the same suggested filename.

The suggested filename will be made unique by appending a number if the
file already exists to support rapid saving of tokens.